:wrench: affair: Separate project into modules
