+++ affair/api/items/usecase/read.py
@@ -1,0 +1,7 @@
+from typing import Optional
+
+def read_item(item_id: int, q: Optional[str] = None):
+    return {
+        "item_id": item_id,
+        "q": q
+    }
