+++ affair/core/models/item.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class Item(BaseModel):
+    name: str
+    price: float
+    is_offer: Optional[bool] = None
