+++ affair/main.py
@@ -1,31 +1,9 @@
-from typing import Optional
-
 from fastapi import FastAPI
-from pydantic import BaseModel
+
+from api.items.usecase.update import update_item
+from api.items.usecase.read import read_item
 
 app = FastAPI()
 
-class Item(BaseModel):
-    name: str
-    price: float
-    is_offer: Optional[bool] = None
-
-@app.get("/")
-def read_root():
-    return {
-        "Hello": "World"
-    }
-
-@app.get("/items/{item_id}")
-def read_item(item_id: int, q: Optional[str] = None):
-    return {
-        "item_id": item_id,
-        "q": q
-    }
-
-@app.put("/items/{item_id}")
-def update_item(item_id: int, item: Item):
-    return {
-        "item_name": item.name,
-        "item_id": item_id
-    }
+app.add_api_route("/items/{item_id}", update_item)
+app.add_api_route("/items/{item_id}", read_item)
